Updated the about section and general section style
