+++ public/scss/components/_about.scss
@@ -1,30 +1,30 @@
 @use '../custom' as *;
-.about {
-    position: relative;
-    //pt-2 pb-2 mt-0 mb-3
-    padding: 20px 0;
-    margin-top: 0;
-    margin-bottom: 30px;
-    .section_content {
-        display: flex;
-        align-items: center;
-        width: 900px;
-        margin: 0 auto;
-        max-width: 80%;
-        &div {
-            margin-right: 25px;
-            &:last-of-type {
-                margin-right: 0;
-            }
-        }
-        // justify-content: center;
-        .pic {
-            width: 800px;
-            max-width: 80%;
-            img {
-                max-width: 100%;
-                height: auto;
-            }
-        }
-    }
-}
+// .about {
+//     position: relative;
+//     //pt-2 pb-2 mt-0 mb-3
+//     padding: 20px 0;
+//     margin-top: 0;
+//     margin-bottom: 30px;
+//     .section_content {
+//         display: flex;
+//         align-items: center;
+//         width: 900px;
+//         margin: 0 auto;
+//         max-width: 80%;
+//         &div {
+//             margin-right: 25px;
+//             &:last-of-type {
+//                 margin-right: 0;
+//             }
+//         }
+//         // justify-content: center;
+//         .pic {
+//             width: 800px;
+//             max-width: 80%;
+//             img {
+//                 max-width: 100%;
+//                 height: auto;
+//             }
+//         }
+//     }
+// }
