+++ public/scss/utils/_typography.scss
@@ -1,4 +1,4 @@
-@use '../custom' as *;
+@use './variables' as *;
 @use './mixins' as mixin;
 
 *{
@@ -70,10 +70,10 @@ footer {
         display: block;
         margin-top: 1.2rem;
         text-transform: none;
-        @include media-breakpoint-up(lg) {
-            font-size: 1.4rem;
-            display: inline-block;
-        }
+        // @include media-breakpoint-up(lg) {
+        //     font-size: 1.4rem;
+        //     display: inline-block;
+        // }
     }
 }
 
@@ -90,9 +90,9 @@ footer {
         background-clip: text;
         -webkit-background-clip: text;
         -webkit-text-fill-color: transparent;
-        @include media-breakpoint-up(lg) {
-            font-size: 1.8rem;
-        }
+        // @include media-breakpoint-up(lg) {
+        //     font-size: 1.8rem;
+        // }
     }
 }
 
@@ -122,42 +122,55 @@ footer {
 // BORDER RIGHT 
 .border-right {
     border-right: 0;
-    @include media-breakpoint-up(sm) {
-        border-right: 0.1rem solid $gray-300;
-        height: 100%;
-    }
+    // @include media-breakpoint-up(sm) {
+    //     border-right: 0.1rem solid $gray-300;
+    //     height: 100%;
+    // }
 }
 
 // GRADIENT BG 
 .gradient {
-    background: $primary; //linear-gradient(to right, rgba($primary, 0.95), rgba($secondary, 0.95));
+    background: $primary;
 }
 
 // SECTIONS
 section {
+    // margin-bottom: 100px;
     margin-top: 100px;
-    margin-bottom: 100px;
-    padding: 150px 0 50px 0 !important;
+    .anchor{
+        height: 150px;
+    }
+        &:first-of-type {
+            margin-top: 0;
+            .anchor {
+                height: 100px;
+            }
+        }
     .section_content {
-        margin-top: 30px;
+        margin-top: 50px;
         max-width: 100%;
     }
-    // &.contact {
-    //     margin: 0;
-    //     .section_content {
-    //         margin: 0px;
-    //         max-width: 100%;
-    //     }
-    // }
-}
 
-header {
-    section {
-        margin: 0;
-        .section_content {
-            margin: 0px;
-            max-width: 100%;
+    &.splitted{
+
+        .section_content{
+            display: grid;
+            grid-template-columns: auto auto;
+
+            position: relative;
+            
+            &::after {
+                content: "";
+                position: absolute;
+                top: -10%;
+                left: 50%;
+                width: 1px;
+                background-color: $gray-200;
+                height: 150%;
+            }
         }
+        
+        
     }
 }
 
@@ -174,5 +187,32 @@ footer {
     }
     .footer_bottom {
         background: transparent;
+    }
+}
+
+
+@media (max-width:760px) {
+// SECTIONS
+    section {
+        margin-top: 10px;
+        .anchor {
+                height: 90px;
+            }
+        &.splitted {
+
+            .section_content {
+                // display: block;
+                grid-template-columns: auto;
+
+                row-gap: 25px;
+                
+
+                &::after {
+                    display: none;
+                }
+            }
+
+
+        }
     }
 }
