+++ src/App.js
@@ -3,7 +3,7 @@ import Header from './components/Header';
 import Footer from './components/Footer';
 import BackToTop from './widgets/back2top';
 import Contact from './components/contact';
-import About from './components/about';
+import About from './components/About';
 import Education from './components/education';
 import Portfolio from './components/Portfolio';
 import MyStacks from './components/mystacks';
