+++ src/components/contact.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Section from '../widgets/section';
+import Section from '../widgets/Section';
 
 const Contact = () => {
     return (
