+++ src/components/mystacks.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Section from '../widgets/section';
+import Section from '../widgets/Section';
 import TeachStack from '../widgets/techstack';
 import { Myinfo } from '../myinfo';
