+++ src/index.js
@@ -1,9 +1,8 @@
 import React,{StrictMode} from 'react';
 import { createRoot } from 'react-dom/client';
-// import ReactDOM from 'react-dom';
-// import './index.css';
+
 import App from './App';
-// import reportWebVitals from './reportWebVitals';
+
 
 const appContainer = document.getElementById('root');
 const root = createRoot(appContainer);
@@ -13,8 +12,3 @@ root.render(
     </StrictMode>
 )
 
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();
